RFC/PR: Require all inline code blocks to be standalone
To make automated testing easier, it should be required for every code block in the API documentation to be run in executors successfully. No code block should require previous code to exist to run an example as a test case.
This would go hand-in-hand with the strict Luau syntax requirements of function declarations.

Existing test cases should retroactively be fixed up to adhere to this new rule.